Move classes into packages/Remove unused imports and classes
