+++ app/src/main/java/com/dankideacentral/dic/activities/OAuthActivity.java
@@ -1,4 +1,4 @@
-package com.dankideacentral.dic.authentication;
+package com.dankideacentral.dic.activities;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
@@ -6,6 +6,7 @@ import android.support.v4.app.FragmentActivity;
 import android.support.v4.app.FragmentTransaction;
 
 import com.dankideacentral.dic.R;
+import com.dankideacentral.dic.fragments.OAuthWebViewFragment;
 
 public class OAuthActivity extends FragmentActivity {
     @Override
