+++ app/src/main/java/com/dankideacentral/dic/activities/SettingsActivity.java
@@ -1,4 +1,4 @@
-package com.dankideacentral.dic;
+package com.dankideacentral.dic.activities;
 
 
 import android.annotation.TargetApi;
@@ -18,6 +18,8 @@ import android.preference.PreferenceManager;
 import android.preference.RingtonePreference;
 import android.text.TextUtils;
 import android.view.MenuItem;
+
+import com.dankideacentral.dic.R;
 
 import java.util.List;
