+++ app/src/main/res/layout/activity_login.xml
@@ -9,7 +9,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.dankideacentral.dic.authentication.LoginActivity">
+    tools:context="com.dankideacentral.dic.activities.LoginActivity">
 
     <ImageView
         android:layout_width="wrap_content"
