+++ app/src/main/res/layout/activity_search.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="com.dankideacentral.dic.SearchActivity">
+    tools:context="com.dankideacentral.dic.activities.SearchActivity">
 
     <android.support.v7.widget.Toolbar
         android:id="@+id/toolbar"
