+++ app/src/main/res/layout/tweet_list.xml
@@ -12,7 +12,7 @@
     android:layout_marginTop="56dp"
     app:layoutManager="LinearLayoutManager"
     android:background="#FFFFFF"
-    tools:context=".TweetListFragment"
+    tools:context=".fragments.TweetListFragment"
     tools:listitem="@layout/tweet_list_item" />
