+++ app/src/main/res/xml/pref_headers.xml
@@ -3,22 +3,22 @@
     <!-- These settings headers are only used on tablets. -->
 
     <header
-        android:fragment="com.dankideacentral.dic.SettingsActivity$AllSettingsFragment"
+        android:fragment="com.dankideacentral.dic.activities.SettingsActivity$AllSettingsFragment"
         android:icon="@drawable/ic_info_black_24dp"
         android:title="All Settings" />
 
     <header
-        android:fragment="com.dankideacentral.dic.SettingsActivity$GeneralPreferenceFragment"
+        android:fragment="com.dankideacentral.dic.activities.SettingsActivity$GeneralPreferenceFragment"
         android:icon="@drawable/ic_info_black_24dp"
         android:title="@string/pref_header_general" />
 
     <header
-        android:fragment="com.dankideacentral.dic.SettingsActivity$NotificationPreferenceFragment"
+        android:fragment="com.dankideacentral.dic.activities.SettingsActivity$NotificationPreferenceFragment"
         android:icon="@drawable/ic_notifications_black_24dp"
         android:title="@string/pref_header_notifications" />
 
     <header
-        android:fragment="com.dankideacentral.dic.SettingsActivity$DataSyncPreferenceFragment"
+        android:fragment="com.dankideacentral.dic.activities.SettingsActivity$DataSyncPreferenceFragment"
         android:icon="@drawable/ic_sync_black_24dp"
         android:title="@string/pref_header_data_sync" />
